# src/python/dart/web/api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /graph/entity_identifiers:
    get:
      operationId: getGraphEntityIdentifiers
      summary: "get the graph entity identifiers"
      tags:
        - "Graph"
        - "GraphEntityIdentifier"
      parameters:
        - name: search
          in: query
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity identifiers"
          schema:
            $ref: '#/definitions/GraphEntityIdentifiersResponse'
        "400":
          description: "error occurred"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /graph/action/{action_id}:
    get:
      operationId: getActionGraphEntity
      summary: "get the action graph entity"
      tags:
        - "Graph"
        - "Action"
      parameters:
        - name: action_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/dataset/{dataset_id}:
    get:
      operationId: getDatasetGraphEntity
      summary: "get the dataset graph entity"
      tags:
        - "Graph"
        - "Dataset"
      parameters:
        - name: dataset_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/datastore/{datastore_id}:
    get:
      operationId: getDatastoreGraphEntity
      summary: "get the datastore graph entity"
      tags:
        - "Graph"
        - "Datastore"
      parameters:
        - name: datastore_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/event/{event_id}:
    get:
      operationId: getEventGraphEntity
      summary: "get the event graph entity"
      tags:
        - "Graph"
        - "Event"
      parameters:
        - name: event_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/subscription/{subscription_id}:
    get:
      operationId: getSubscriptionGraphEntity
      summary: "get the subscription graph entity"
      tags:
        - "Graph"
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/trigger/{trigger_id}:
    get:
      operationId: getTriggerGraphEntity
      summary: "get the trigger graph entity"
      tags:
        - "Graph"
        - "Trigger"
      parameters:
        - name: trigger_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/workflow/{workflow_id}:
    get:
      operationId: getWorkflowGraphEntity
      summary: "get the workflow graph entity"
      tags:
        - "Graph"
        - "Workflow"
      parameters:
        - name: workflow_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity"
          schema:
            $ref: '#/definitions/GraphEntityResponse'
  /graph/sub_graph:
    get:
      operationId: getSubgraphGraphEntity
      summary: "get the sub graphs"
      tags:
        - "Graph"
        - "Subgraph"
      parameters:
        - name: related_type
          in: query
          type: string
        - name: related_id
          in: query
          type: string
        - name: engine_name
          in: query
          type: string
      responses:
        "200":
          description: "returns the sub graphs"
          schema:
            $ref: '#/definitions/ObjectResponse'
        "404":
          description: "error occurred"
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: createSubgraphGraphEntities
      summary: "create a new sub graph"
      tags:
        - "Graph"
      parameters:
        - name: debug
          in: query
          type: boolean
        - name: graph_entities
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/GraphEntity"
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            $ref: '#/definitions/ObjectResponse'
        "400":
          description: "error occurred"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /schema/engine/{engine_name}/datastore:
    get:
      operationId: getEngineDatastoreSchema
      summary: "get the datastore JSON Schema for the named engine"
      tags:
        - "Schema"
        - "Engine"
        - "Datastore"
      parameters:
        - name: engine_name
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            $ref: '#/definitions/JSONSchemaResponse'
        "400":
          description: "if engine is unknown"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /schema/datastore:
    get:
      operationId: getDatastoreSchema
      summary: "get the datastore JSON Schema"
      tags:
        - "Schema"
        - "Datastore"
      parameters:
        - name: datastore_id
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            $ref: '#/definitions/JSONSchemaResponse'
        "404":
          description: "if the optionally specified datastore could not be found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /schema/action:
    get:
      operationId: getActionSchema
      summary: "get the action JSON Schema"
      tags:
        - "Schema"
        - "Action"
      parameters:
        - name: action_id
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            $ref: '#/definitions/JSONSchemaResponse'
        "404":
          description: "if the optionally specified action could not be found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /schema/action/{action_name}:
    get:
      operationId: getActionSchemaByName
      summary: "get the action JSON Schema by action name"
      tags:
        - "Schema"
        - "Action"
      parameters:
        - name: action_name
          in: path
          type: string
          required: true
        - name: action_id
          in: query
          type: string
        - name: datastore_id
          in: query
          type: string
        - name: workflow_id
          in: query
          type: string
        - name: engine_name
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            $ref: '#/definitions/JSONSchemaResponse'
        "400":
          description: "if the optionally specified action is unknown"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "if the optionally specified action could not be found"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /schema/trigger:
    get:
      operationId: getTriggerSchema
      summary: "get the trigger JSON schema"
      tags:
        - "Schema"
        - "Trigger"
      parameters:
        - name: trigger_type_name
          in: query
          type: string
        - name: trigger_id
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            $ref: '#/definitions/JSONSchemaResponse'
        "400":
          description: "if the optionally specified trigger is unknown"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
...
  GraphEntity:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
      name:
        type: string
      related_id:
        type: string
      related_is_a:
        type: string
      related_type:
        type: string
      state:
        type: string
      sub_type:
        type: string
    type: object
  GraphEntityIdentifier:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
      name:
        type: string
    type: object
...
  GraphEntityIdentifiersResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/GraphEntityIdentifier'
        type: array
    type: object
  GraphEntityResponse:
    properties:
      results:
        $ref: '#/definitions/GraphEntity'
    type: object
...
  JSONSchema:
    description: JSON Schema object as defined by http://json-schema.org/schema
    type: object
  JSONSchemaResponse:
    properties:
      results:
        $ref: '#/definitions/JSONSchema'
    type: object
...
  ObjectResponse:
    properties:
      results:
        type: object
    type: object
...